Properly load history lines and stops
Now a single SQL call is made, and such call uses LEFT OUTER JOIN to directly obtain the line or the stop in one go.
Now the architecture is also better because there is no Any passed around.
Now LiveData listens not only on the HistoryEntry table, but also on DbStop and DbLine, which means that the drawer items auto update in case of stop+line reloading.